Move fixed Jacobian helpers to linear internals
