--- gtsam/linear/FixedJacobianFactorOps.cpp
@@ -1,19 +1,0 @@
-/* ----------------------------------------------------------------------------
-
- * GTSAM Copyright 2010-2026, Georgia Tech Research Corporation,
- * Atlanta, Georgia 30332-0415
- * All Rights Reserved
- * Authors: Frank Dellaert, et al. (see THANKS for the full author list)
-
- * See LICENSE for the license information
-
- * -------------------------------------------------------------------------- */
-
-/**
- * @file FixedJacobianFactorOps.cpp
- * @brief Explicit instantiations of common fixed-size Hessian kernels.
- */
-
-#define GTSAM_FIXED_JACOBIAN_DEFINE_INSTANTIATIONS
-#include <gtsam/linear/FixedJacobianFactorOps.h>
-#undef GTSAM_FIXED_JACOBIAN_DEFINE_INSTANTIATIONS
