+++ gtsam/linear/TernaryJacobianFactor.h
@@ -12,13 +12,14 @@
 /**
  * @file TernaryJacobianFactor.h
  * @brief Fixed-size specialization of a three-key JacobianFactor.
+ * @author Frank Dellaert
  */
 
 #pragma once
 
 #include <gtsam/base/timing.h>
-#include <gtsam/linear/FixedJacobianFactorOps.h>
 #include <gtsam/linear/JacobianFactor.h>
+#include <gtsam/linear/internal/FixedJacobianFactorOps.h>
 
 #include <algorithm>
