+++ gtsam_unstable/dynamics/SimpleHelicopter.h
@@ -1,3 +1,18 @@
+/* ----------------------------------------------------------------------------
+
+ * GTSAM Copyright 2010, Georgia Tech Research Corporation,
+ * Atlanta, Georgia 30332-0415
+ * All Rights Reserved
+ * Authors: Frank Dellaert, et al. (see THANKS for the full author list)
+
+ * See LICENSE for the license information
+
+ * -------------------------------------------------------------------------- */
+
+/**
+ * @file SimpleHelicopter.h
+ */
+
 /*
  * @file SimpleHelicopter.h
  * @brief Implement SimpleHelicopter discrete dynamics model and variational integrator,
