+++ gtsam_unstable/slam/LocalOrientedPlane3Factor.h
@@ -1,3 +1,18 @@
+/* ----------------------------------------------------------------------------
+
+ * GTSAM Copyright 2010, Georgia Tech Research Corporation,
+ * Atlanta, Georgia 30332-0415
+ * All Rights Reserved
+ * Authors: Frank Dellaert, et al. (see THANKS for the full author list)
+
+ * See LICENSE for the license information
+
+ * -------------------------------------------------------------------------- */
+
+/**
+ * @file LocalOrientedPlane3Factor.h
+ */
+
 /*
  * @file LocalOrientedPlane3Factor.h
  * @brief LocalOrientedPlane3 Factor class
