+++ gtsam_unstable/slam/RelativeElevationFactor.cpp
@@ -1,3 +1,14 @@
+/* ----------------------------------------------------------------------------
+
+ * GTSAM Copyright 2010, Georgia Tech Research Corporation,
+ * Atlanta, Georgia 30332-0415
+ * All Rights Reserved
+ * Authors: Frank Dellaert, et al. (see THANKS for the full author list)
+
+ * See LICENSE for the license information
+
+ * -------------------------------------------------------------------------- */
+
 /**
  * @file RelativeElevationFactor.cpp
  *
@@ -50,4 +61,3 @@ void RelativeElevationFactor::print(const std::string& s, const KeyFormatter& ke
 /* ************************************************************************* */
 
 } // \namespace gtsam
-
