+++ gtsam_unstable/slam/RelativeElevationFactor.h
@@ -1,3 +1,14 @@
+/* ----------------------------------------------------------------------------
+
+ * GTSAM Copyright 2010, Georgia Tech Research Corporation,
+ * Atlanta, Georgia 30332-0415
+ * All Rights Reserved
+ * Authors: Frank Dellaert, et al. (see THANKS for the full author list)
+
+ * See LICENSE for the license information
+
+ * -------------------------------------------------------------------------- */
+
 /**
  * @file RelativeElevationFactor.h
  *
@@ -83,4 +94,3 @@ private:
 
 
 } // \namespace gtsam
-
